src/ml/requirements_ml.txt
# Übung 5 - ML Requirements
# Core data processing
pandas>=1.5.0
numpy>=1.20.0

# Machine Learning - Professor's Recommendations
darts>=0.24.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Explainable AI
shap>=0.41.0

# Class imbalance handling
imbalanced-learn>=0.9.0

# Database connectivity (already installed)
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
